[#20] 캐릭터 클래스 내용 및 구조 변경

+++ Source/Y2024Q1_Prototyping/Private/UI/Player/LLL_PlayerStatusWidget.cpp
@@ -12,7 +12,7 @@ void ULLL_PlayerStatusWidget::NativeConstruct()
 	
 }
 
-void ULLL_PlayerStatusWidget::UpdateWidgetView(int MaxHealth, int CurrentHealth, int MaxShield, int CurrentShield) const
+void ULLL_PlayerStatusWidget::UpdateWidgetView(float MaxHealth, float CurrentHealth, float MaxShield, float CurrentShield) const
 {
 	if(MaxHealth)
 	{
